bin/ci/requirements.txt
colorama==0.4.5
filecache==0.81
gql==3.4.0
python-dateutil==2.8.2
pandas==2.1.1
plotly==5.17.0
python-gitlab==3.5.0
PyYAML==6.0.1
ruamel.yaml.clib==0.2.8
ruamel.yaml==0.17.21
